requirements.txt
discord.py==1.2.4
wikipedia==1.4.0
redis==3.3.11
google-images-download==2.8.0